requirements.txt
# Core
numpy==1.24.3
opencv-python==4.8.1.78
mediapipe==0.10.12
pygame==2.6.1
pillow==10.0.0
matplotlib==3.8.0
tk==0.1.0

# TensorFlow for CNN
tensorflow==2.13.0

# SMS/Email alerts (example: Twilio for SMS, smtplib is built-in)
twilio==8.14.0

# Optional: for sending emails using Gmail via secure SMTP
email-validator==2.1.0